move user-facing functions to ctarget folder
