+++ core/tag.c
@@ -94,24 +94,8 @@ instant_t _lf_last_reported_unadjusted_physical_time_ns = NEVER;
 /**
  * Return the current tag, a logical time, microstep pair.
  */
-tag_t lf_tag() {
+tag_t _lf_tag() {
     return current_tag;
-}
-
-/**
- * Return the current tag, a logical time, microstep pair.
- * @deprecated
- */
-tag_t get_current_tag() {
-    return lf_tag();
-}
-
-/**
- * Return the current microstep.
- * @deprecated
- */
-microstep_t get_microstep() {
-    return lf_tag().microstep;
 }
 
 /**
@@ -124,7 +108,7 @@ microstep_t get_microstep() {
  * @param tag2
  * @return -1, 0, or 1 depending on the relation.
  */
-int lf_tag_compare(tag_t tag1, tag_t tag2) {
+int _lf_tag_compare(tag_t tag1, tag_t tag2) {
     if (tag1.time < tag2.time) {
         return -1;
     } else if (tag1.time > tag2.time) {
@@ -136,13 +120,6 @@ int lf_tag_compare(tag_t tag1, tag_t tag2) {
     } else {
         return 0;
     }
-}
-
-/**
- * @deprecated version of 'lf_tag_compare'
- */
-int compare_tags(tag_t tag1, tag_t tag2) {
-    return lf_tag_compare(tag1, tag2);
 }
 
 /**
@@ -180,29 +157,6 @@ tag_t _lf_delay_tag(tag_t tag, interval_t interval) {
     }
     return result;
 }
-
-/**
- * Return the current logical time in nanoseconds since January 1, 1970.
- */
-instant_t lf_time_logical() {
-    return _lf_time(LF_LOGICAL);
-}
-/**
- * @deprecated version of "lf_time_logical"
- */
-instant_t get_logical_time() { return lf_time_logical(); }
-
-/**
- * Return the elapsed logical time in nanoseconds since the start of execution.
- */
-interval_t lf_time_logical_elapsed() {
-    return _lf_time(LF_ELAPSED_LOGICAL);
-}
-/**
- * @deprecated version of "lf_time_logical_elapsed"
- */
-interval_t get_elapsed_logical_time() { return lf_time_logical_elapsed(); }
-
 
 /**
  * Return the current physical time in nanoseconds since January 1, 1970,
@@ -247,47 +201,6 @@ instant_t _lf_physical_time() {
 }
 
 /**
- * Return the current physical time in nanoseconds since January 1, 1970,
- * adjusted by the global physical time offset.
- */
-instant_t lf_time_physical() {
-    return _lf_time(LF_PHYSICAL);
-}
-/**
- * @deprecated version of "lf_time_physical"
- */
-instant_t get_physical_time() { return lf_time_physical(); }
-
-/**
- * Return the elapsed physical time in nanoseconds.
- * This is the time returned by get_physical_time() minus the
- * physical start time as measured by get_physical_time() when
- * the program was started.
- */
-instant_t lf_time_physical_elapsed() {
-    return _lf_time(LF_ELAPSED_PHYSICAL);
-}
-/**
- * @deprecated version of "lf_time_physical_elapsed"
- */
-instant_t get_elapsed_physical_time() { return lf_time_physical_elapsed(); }
-
-/**
- * Return the physical time of the start of execution in nanoseconds. * 
- * On many platforms, this is the number of nanoseconds
- * since January 1, 1970, but it is actually platform dependent. * 
- * @return A time instant.
- */
-instant_t lf_time_start(void) {
-    return _lf_time(LF_START);
-}
-/**
- * @deprecated version of "lf_time_start"
- */
-instant_t get_start_time() { return lf_time_start(); }
-
-
-/**
  * Get the time specified by "type".
  * 
  * Example use cases:
@@ -328,15 +241,8 @@ instant_t _lf_time(_lf_time_type type) {
  * and get_elpased_physical_time() will have this specified offset
  * added to what it would have returned before the call.
  */
-void lf_set_physical_clock_offset(interval_t offset) {
+void _lf_set_physical_clock_offset(interval_t offset) {
     _lf_time_test_physical_clock_offset += offset;
-}
-
-/**
- * @deprecated version of 'lf_set_physical_clock_offset'
- */
-void set_physical_clock_offset(interval_t offset) {
-    lf_set_physical_clock_offset(offset);
 }
 
 /**
