+++ core/tag.h
@@ -112,6 +112,11 @@ typedef struct {
 typedef tag_t tag_interval_t;
 
 /**
+ * Return the current tag, a logical time, microstep pair.
+ */
+tag_t _lf_tag();
+
+/**
  * Compare two tags. Return -1 if the first is less than
  * the second, 0 if they are equal, and +1 if the first is
  * greater than the second. A tag is greater than another if
@@ -121,24 +126,7 @@ typedef tag_t tag_interval_t;
  * @param tag2
  * @return -1, 0, or 1 depending on the relation.
  */
-int lf_tag_compare(tag_t tag1, tag_t tag2);
-DEPRECATED(int compare_tags(tag_t tag1, tag_t tag2));
-
-/**
- * Return the current tag, a logical time, microstep pair.
- */
-tag_t lf_tag();
-
-/**
- * Return the current tag, a logical time, microstep pair.
- */
-DEPRECATED(tag_t get_current_tag(void));
-
-/**
- * Return the current microstep.
- */
-DEPRECATED(microstep_t get_microstep(void));
-
+int _lf_tag_compare(tag_t tag1, tag_t tag2);
 
 /**
  * Store into the specified buffer a string giving a human-readable
@@ -198,58 +186,12 @@ typedef enum _lf_time_type {
 instant_t _lf_time(_lf_time_type type);
 
 /**
- * Return the current logical time in nanoseconds.
- * On many platforms, this is the number of nanoseconds
- * since January 1, 1970, but it is actually platform dependent.
- * 
- * @return A time instant.
- */
-instant_t lf_time_logical(void);
-DEPRECATED(instant_t get_logical_time(void));
-
-/**
- * Return the elapsed logical time in nanoseconds
- * since the start of execution.
- * @return A time interval.
- */
-interval_t lf_time_logical_elapsed(void);
-DEPRECATED(interval_t get_elapsed_logical_time(void));
-
-/**
- * Return the current physical time in nanoseconds.
- * On many platforms, this is the number of nanoseconds
- * since January 1, 1970, but it is actually platform dependent.
- * @return A time instant.
- */
-instant_t lf_time_physical(void);
-DEPRECATED(instant_t get_physical_time(void));
-
-/**
- * Return the elapsed physical time in nanoseconds.
- * This is the time returned by get_physical_time(void) minus the
- * physical start time as measured by get_physical_time(void) when
- * the program was started.
- */
-instant_t lf_time_physical_elapsed(void);
-DEPRECATED(instant_t get_elapsed_physical_time(void));
-
-/**
- * Return the physical and logical time of the start of execution in nanoseconds.
- * On many platforms, this is the number of nanoseconds
- * since January 1, 1970, but it is actually platform dependent. 
- * @return A time instant.
- */
-instant_t lf_time_start(void);
-DEPRECATED(instant_t get_start_time(void));
-
-/**
  * Set a fixed offset to the physical clock.
- * After calling this, the value returned by get_physical_time(void)
- * and get_elpased_physical_time(void) will have this specified offset
+ * After calling this, the value returned by get_physical_time()
+ * and get_elpased_physical_time() will have this specified offset
  * added to what it would have returned before the call.
  */
-void lf_set_physical_clock_offset(interval_t offset);
-DEPRECATED(void set_physical_clock_offset(interval_t offset));
+void _lf_set_physical_clock_offset(interval_t offset);
 
 /**
  * Delay a tag by the specified time interval to realize the "after" keyword.
